Moved item-action-controls into a sub-directory
